Check $userChangedTo when determining 'no-cache'.
Update class Listener function emptyCacheForLoggedUser() to now also
check the value of self::$userChangedTo when determining if the
'no-cache' header should be sent or not.

This change should resolve a reversion issue where visitors logged in
through the "forgot my password" form submission would have their
logged-in cookie set and then unintentionally unset because XenForo's
Visitor class object is not updated to reflect that the user will now be
logged in, only the User object is ever updated.

upload/src/addons/LiteSpeedCache/Listener.php
[Update] emptyCacheForLoggedUser(): Added a check on
self::$userChangedTo when determining whether or not to set the
'no-cache' header.